The National Health and Nutrition Examination Survey (NHANES)
The National Health and Nutrition Examination Survey (NHANES) is a nationally representative health study conducted by the CDC’s National Center for Health Statistics. It uniquely combines interviews, physical exams, and lab tests to assess the health and nutritional status of U.S. children and adults. 

 

This dataset is a subset, focused part of the NHANES study by CDC, which looks at health and nutrition trends in the U.S. It has 6,287 entries and 7 key features: covering things like body stats, lifestyle, and lab results. The task is simple: predict if a person is a senior (65+) or not. NHANES collected data through home visits, mobile clinics, and lab tests, giving a good mix of real and reported info. This trimmed version keeps only what’s needed, making it perfect for health-based age prediction.

 

This Hackathon is in collaboration with by Consulting and Analytics Club, IIT Guwahati for Summer Analytics 2025.